Add AISegment - UNet model reduced test npy file
Sample numpy file is added for AISegment dataset that has shape
(48x48x48) and is the folded version of an (192x192x3) input image.

AISegment yaml file and gen-demos script's related parts are modified
accordingly.

With this sample numpy file, demo can be successfully generated and KAT
can be run on EVKit or Feather board.