Add sampleoutput.h parser
Sample notebook for parsing izer generated ampleoutput.h file.

Sample numpy input is fed to the PyTorch model that is read from
training repository and values are compared with the memeory cotents
logged in the sampleoutput.h file.

Sample notebook is designed for verificaiton of SVHN dataset and
tinier-ssd model.